WIP #Adv _getline
the simple version if now the blueprint for this one to be rewriten in.

Add a work in progress getline() function.
currently functions (excluding `_getline()` )are static to the source, and Will later be broken out
into the header once it's passing functioning.

This version is currently not functions all and it going to be heavely rewriten.